Split SvgElement
